Revert "refactor(jsonl): keep legacy exporter in export file"
This reverts commit 2610e2153ec130a0c5ecb479e2a0873ef27739ee.
